Added method FindFreeTime
FindFreeTime finds when is doctor available for appointment (operation or examination). Also saved some tests for this method.
Added Operation.cs and Examination.cs.

+++ HealthInstitution/MVVM/Models/Entities/AppointmentReview.cs
@@ -8,17 +8,17 @@ namespace HealthInstitution.MVVM.Models.Entities
 {
     public class AppointmentReview
     {
-        private int _rating;
+        private double _rating;
         private string _comment;
 
-        public AppointmentReview(int rating, string comment)
+        public AppointmentReview(double rating, string comment)
         {
             _rating = rating;
             _comment = comment;
         }
 
-        public int GetRating() => _rating;
-        public void SetRating(int rating) => _rating = rating;
+        public double GetRating() => _rating;
+        public void SetRating(double rating) => _rating = rating;
         public string GetComment() => _comment;
         public void SetComment(string comment) => _comment = comment;
     }
